requirements/base.txt
Django==4.2.5
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-filter==23.2
djangorestframework==3.14.0
djoser==2.2.0
drf-yasg==1.21.7
pre-commit==3.4.0
python-dotenv==1.0.0
